autoxloo-ai-assistant/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/ML
crewai==0.1.26
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
pinecone-client==3.0.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23

# Utilities
faker==20.1.0
httpx==0.25.2
